Add 'shell_quoting' feature to FileChooserEntry.
This is used for the prelaunch and postlaunch commands. When on, the entry has distinct text and path values - text is what the entry shows and the 'real value' but the path will be extracted from the text via shlex.split().

You can then browse for an set the path, leaving any other arguments alone.

This also ensure that if that path needs quoting, it gets quoting.

In most cases the 'text' and 'path' are identical, which should keep existing code happy.

Resolves #4989