Fix launching games from tray icon
The tray icon class calls Application.get_launch_ui_delegate, which is not defined in Application.
I wasn't sure if it was better to make the icon class use directly Application.launch_ui_delegate or add the getter method. I opted for the getter method so that the field can be treated as an implementation detail.