Do editable-grids (environment vars) in options with boxes, not Gtk.Grid.
Gtk.Grid is madness on toast, and confuses not just be but GTK's layout engine, which bugs out and places this wrongly when you filter the options. It also mis-sizes the buttons under the grid.

Replace the Gtk.Grid with a couple of Gtk.Box. Caveman brain know Box. Also, size-requests for things.

I've sized the editable grid so that it looks like a nice even number of rows for me. Heavy knows what will happen when your favorite theme though.