Move Runtime away from None returns
Failures to access files and directories can be exceptions, with try-catch instead of if-then.

get_updated_at is obvious; just catch the FileNotFoundError.

local_runtime_path is a bit less so; we want to raise a ValueError early, when the Runtime is constructed, so we don't wind up with a broken Runtime object, rather than coping in each method.

We still have Nones that pass through an executor. I'll need to research that.