Fix executor error handling
So, the return-None-on-error code was broken. The actual result of download_component was never checked - it was looking at the filename listed in the component dict, which is not going to be None even if there is an error.

The fix is to let exception propagate to the loop over the futures and check for them there. Same behavior, simple type annotation, and it works now.