Re-introduce old remove_folder as delete_folder
There are several places where we need the directory deleted now, not asynchronously, because we are about to replace it. I think these are in places where deletion is okay - stuff like installing runtime updates.

Also, I put deletion of temp directories back on the deletion plan. I don't think anyone wants to see our temp files in the trash.