Add plain old try-except handling for style_manager.py
The handlers log the error and continue; this prevents Lutris from crashing on startup, but leaves you insensitive to the system-wide dark-theme-light-theme setting.

We already handle errors this way for the TrashPortal.

I suspect the right answer here is to wrap DBus in some more civilized interface that could take care of this. But this commit will help until we figure that out.