Add logic to ensure destruction of ModalDialogs
We never re-use these but often forget to destroy them. So, this commit schedules an idle-function to destroy them as long as you don't reshow them before them, or destroy them yourself.

The main thing is to defer this until after run() has returned- destroy() resets the result from that function, if it happens first.