Make Uninstall Game dialog modeless
The trick here is being able to add games to it, if you open it then go back to the LutrisWindow and remove some more games.

Doing this in turn can cause games already listed to become 'more deletable' since the game-directory that was shared with another game may now be deletable- if you are deleting all the relevant games.

At least this breaks up the big __init__ method we had here.