Rework the way extra files are passed through to the installer script.
Instead of passing the extra-ids everwhere and hoping it lines up, we pass them to the installer to generate the InstallerFiles, but it provides the extra files separately.

We can then extract the download destination paths for each extra file we're downloading, and we can put those in the script.

This way, if there are more or fewer files than we expect, it will still work.

This may help with #4974, but it's hard to be sure.