Add some error handling for the error handling!
It's very easy to pass the error to get_error_handler(), but we want the type signature to be simple, not Union[TError, Type[TError]], so instead we'll log if you pass the exception and fix it at runtime.

We will raise another exception if you pass some random thing in as well, since the 'attribute __mro__ not found' is just confusing.