Rework get_default_runner_version_info() to avoid failure when 'versions.json' is stale.
Even if it is stale, even whe a specific version is requested, we'll try to satisfy version info request from the cache and hit the API only if this fails.

This commit also breaks out the code to parse out the architecture from the version, and uses this in the runner-version dialog too.

 That bit uses Optional[str], which is not ideal, but I haven't figured out a better way here. Just using '' instead is cheating, a None spelled differently.