Replace game filter with more efficient version.
This will just leave the list alone if 'Installed games only' is off and the filter text is empty.

Also, it no longer strips the filter text over and over again, but just once at the start. It still has to strip each game name, but it's half the strippings now.

Also, fix some ruff issues.