Fix NoneType (and other typing errors) in GOG service code.
Added lots of type annotations to find some of these.

Also, moved some exception handling to a higher point. We'll still try to load extras without DLC if the DLC isn't accessible, but in other contexts failure to reach GOG will let an exception out, to be reported to the user.

What it won't do it turn these into None returns.