Replace dark_theme with light_theme
This preference forces Lutris over to light-mode, instead of to dark mode, since the default is now dark mode.

I find that in GNOME the "default" is NO_PREFERENCE, which is now be treated as dark mode. This will override it to be light, but a system that actual reports color-scheme 2 (PREFER_LIGHT) will always be light mode, with no override possible anymore.