Remove executable bit from desktop entries
Desktop entries don't need to be executable since they are not actually
executed by the application launcher, but rather metadata for an
installed binary.

Executing a desktop files also merely results in a lot of junk output on
the terminal.

Additionally, any permission that is not required should not be granted,
see Principle of least Privilege.